Finish moving all endpoints to memoized fns
