+++ src/lib/utils.js
@@ -1,13 +1,17 @@
-export async function fetch_page(url, params = {}) {
-	try {
-		let response = await fetch(url, params);
+import pLimit from 'p-limit';
+const limit = pLimit(2);
 
-		if (!response.ok) {
+export const fetch_page = async (url, params = {}) =>
+	limit(async () => {
+		try {
+			let response = await fetch(url, params);
+
+			if (!response.ok) {
+				throw new Error('Unable to fetch');
+			}
+
+			return response;
+		} catch {
 			throw new Error('Unable to fetch');
 		}
-
-		return response;
-	} catch {
-		throw new Error('Unable to fetch');
-	}
-}
+	});
