+++ src/routes/+layout.server.js
@@ -1,8 +1,8 @@
-import cinemas from '$lib/api/cinemas.js';
+import get_cinemas from '$lib/api/cinemas.js';
 
 export const prerender = true;
 export async function load() {
 	return {
-		cinemas: await cinemas()
+		cinemas: await get_cinemas()
 	};
 }
