+++ src/routes/cinema/[cinema_id]/+page.server.js
@@ -1,32 +1,18 @@
-import pLimit from 'p-limit';
-import pRetry from 'p-retry';
 import get_movie from '$lib/api/movie.js';
-const limit = pLimit(2);
+import get_movies from '$lib/api/movies.js';
+import get_rating from '$lib/api/rating.js';
+import get_times from '$lib/api/times.js';
 
-const fetch_with_retry = async (fetch_fn, url) => {
-	let response = await fetch_fn(url);
-	if (!response.ok) {
-		throw new Error(response.statusText);
-	}
-	return response.json();
-};
-
-export async function load({ fetch, params }) {
-	console.log(`Begin fetching ${params.cinema_id}...`);
+export async function load({ params }) {
+	console.log(`Fetching data for cinema with id ${params.cinema_id}...`);
 
 	// Get master list of movies
-	let movielist = await pRetry(() => fetch_with_retry(fetch, '/api/movies'));
+	let movielist = await get_movies();
 
 	// Get times for each movie
 	let times = await Promise.all(
-		movielist.movies.map((movie) =>
-			limit(() =>
-				pRetry(() => fetch_with_retry(fetch, `/api/times/${movie.id}/${params.cinema_id}`))
-			)
-		)
+		movielist.movies.map((movie) => get_times(movie.id, params.cinema_id))
 	);
-
-	console.log(`Got all times for ${params.cinema_id}...`);
 
 	// Add times to the object and filter out movies without any playtimes
 	movielist = movielist.movies.reduce((prev, cur, i) => {
@@ -41,18 +27,10 @@ export async function load({ fetch, params }) {
 	}, []);
 
 	// Get pathe info (including year)
-	let info = await Promise.all(
-		movielist.map((movie) =>
-			limit(() => get_movie(movie.id))
-		)
-	);
+	let info = await Promise.all(movielist.map((movie) => get_movie(movie.id)));
 
 	// Get RT ratings
-	let ratings = await Promise.all(
-		movielist.map((movie) =>
-			limit(() => pRetry(() => fetch_with_retry(fetch, `/api/rating/${movie.id}`)))
-		)
-	);
+	let ratings = await Promise.all(movielist.map((movie) => get_rating(movie.id)));
 
 	// Combine object
 	movielist = movielist.map((movie, i) => {
@@ -60,8 +38,6 @@ export async function load({ fetch, params }) {
 		movie.rt = ratings[i];
 		return movie;
 	});
-
-	console.log(`Done fetching ${params.cinema_id}...`);
 
 	return {
 		movies: movielist
